+++ packages/flutter_state_notifier/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update dependencies
 - Improved the error when a listener of a `StateNotifier` throws to include
   the thrown error/stacktrace
+- `StateNotifier.state =` now filters updates using `identical`.
 
 ## 0.7.0